Feat Added pick objects package
Created the pick objects package for the project.
It's used to control navigation of the robot to goal states.
It picks the objects that need retrieving and send those coordinates to the robot
Once the object is picked up, it gives the home base coordintaes to the robot to take the object to