Fix Fixed rviz config directory and marker scale
Moved the rvizConfig directory into the my_robot package so it can be found in a search
Updated the launch file to reference it at it's new location

Set the scale of the marker. By default it's 0. It need to have a size to be seen.

Set the goal in pick objects to match the add_marker coordinates

Made a new bash script to launch everything for the service robot project.